Extract export handler creator hof
