+++ export/api/export/image.ts
@@ -1,87 +1,47 @@
-import * as Sentry from '@sentry/node';
-import type { NowApiHandler, NowRequest, NowResponse } from '@vercel/node';
-import type { Page } from 'puppeteer';
 import _ from 'lodash';
 
-import * as render from '../../src/render-serverless';
-import config from '../../src/config';
 import { validateExportData } from '../../src/data';
-import { render422, render500 } from '../../src/views';
-import { HttpError } from '../../src/HttpError';
-import { setUpSentry, throwIfAcademicYearNotSet } from '../../src/serverless-helpers';
+import { makeExportHandler } from '../../src/handler';
+import * as render from '../../src/render-serverless';
+import type { PageData } from '../../src/types';
 
-const handler: NowApiHandler = async (request, response) => {
-  try {
-    await main(request, response);
-  } catch (e) {
-    const eventId = Sentry.captureException(e.original || e);
-
-    console.error(e);
-
-    if (e instanceof HttpError) {
-      if (e.code === 422) {
-        response.status(422).send(render422());
-        return;
-      }
-    }
-    response.status(500).send(render500(eventId));
-  }
+type Data = {
+  exportData: PageData;
+  options: render.ViewportOptions;
 };
 
-async function main(request: NowRequest, response: NowResponse) {
-  throwIfAcademicYearNotSet();
-  setUpSentry();
+const handler = makeExportHandler<Data>(
+  (request) => {
+    const exportData = JSON.parse(request.query.data as never);
+    validateExportData(exportData);
 
-  // Validate input before starting the browser (which is expensive)
-  let data = undefined;
-  try {
-    const dataCandidate = JSON.parse(request.query.data as never);
-    validateExportData(dataCandidate);
-    data = dataCandidate;
-  } catch (e) {
-    throw new HttpError(422, 'Invalid timetable data', e);
-  }
-
-  let options: render.ViewportOptions = {
-    pixelRatio: _.clamp(Number(request.query.pixelRatio) || 1, 1, 3),
-  };
-  const height = Number(request.query.height);
-  const width = Number(request.query.width);
-  if (
-    typeof height !== 'undefined' &&
-    typeof width !== 'undefined' &&
-    !Number.isNaN(height) && // accept floats
-    !Number.isNaN(width) && // accept floats
-    height > 0 &&
-    width > 0
-  ) {
-    options = { ...options, height, width };
-  }
-
-  // Prepare browser for export
-  const url = config.page;
-  let page: Page;
-  try {
-    page = await render.open(url);
-  } catch (e) {
-    if (e.message.includes('ERR_CONNECTION_REFUSED')) {
-      throw new HttpError(
-        500,
-        `Could not open the page located at process.env.PAGE (${url}). Try opening it in your browser?`,
-        e,
-      );
+    let options: render.ViewportOptions = {
+      pixelRatio: _.clamp(Number(request.query.pixelRatio) || 1, 1, 3),
+    };
+    const height = Number(request.query.height);
+    const width = Number(request.query.width);
+    if (
+      typeof height !== 'undefined' &&
+      typeof width !== 'undefined' &&
+      !Number.isNaN(height) && // accept floats
+      !Number.isNaN(width) && // accept floats
+      height > 0 &&
+      width > 0
+    ) {
+      options = { ...options, height, width };
     }
-    throw new HttpError(500, 'Cannot start browser', e);
-  }
 
-  // Export
-  const body = await render.image(page, data, options);
-  response.setHeader('Content-Disposition', 'attachment; filename="My Timetable.png"');
-  response.setHeader('Content-Type', 'image/png');
-  response.status(200).send(body);
-
-  // Cleanup
-  await page.close();
-}
+    return {
+      exportData,
+      options,
+    };
+  },
+  async (response, page, { exportData, options }) => {
+    const body = await render.image(page, exportData, options);
+    response.setHeader('Content-Disposition', 'attachment; filename="My Timetable.png"');
+    response.setHeader('Content-Type', 'image/png');
+    response.status(200).send(body);
+  },
+);
 
 export default handler;
